Refactor main.js into modules
