+++ index.html
@@ -227,6 +227,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js/main.js" defer></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
